Add ability for non-MOSS LOTW to use the element token pool
Closes #1521
