+++ CauldronMods/DeckLists/Hero/LadyOfTheWoodDeckList.json
@@ -42,7 +42,17 @@
                 "CeladrochCharacter": "I haven't felt this presence in many years. Beware, my friends. This wind carries the scent of death.",
                 "HydraWinterTiamatCharacter": "If we do not stop her, she will bring ruin upon all these lands!"
             },
-            "complexity": 3
+            "complexity": 3,
+            "tokenPools": [
+                {
+                    "identifier": "LadyOfTheWoodElementPool",
+                    "name": "Lady of the Wood's Element Pool",
+                    "initialValue": 0,
+                    "minimumValue": 0,
+                    "labelTopColor": "e04b26",
+                    "labelBottomColor": "e04b26"
+                }
+            ]
         },
         {
             "identifier": "CalmBeforeTheStorm",
@@ -485,7 +495,17 @@
                 "CeladrochCharacter": "Hurry, everyone! His bonds are weakening - but ours can strengthen!",
                 "HydraWinterTiamatCharacter": "Poor beast. She has become an abomination, unable to rest, driven by cruelty."
             },
-            "complexity": 3
+            "complexity": 3,
+            "tokenPools": [
+                {
+                    "identifier": "LadyOfTheWoodElementPool",
+                    "name": "Lady of the Wood's Element Pool",
+                    "initialValue": 0,
+                    "minimumValue": 0,
+                    "labelTopColor": "e04b26",
+                    "labelBottomColor": "e04b26"
+                }
+            ]
         },
         {
             "identifier": "LadyOfTheWoodCharacter",
@@ -529,7 +549,7 @@
                 "HydraWinterTiamatCharacter": "Come, friends. We must deal with a dragon."
             },
             "complexity": 3,
-             "tokenPools":[
+            "tokenPools":[
                 {
                     "identifier": "LadyOfTheWoodElementPool",
                     "name": "Lady of the Wood's Element Pool",
@@ -577,9 +597,19 @@
                 "default": "I have seen millenia slip away; as will your chance at victory.",
                 "CeladrochCharacter": "It seems like not all things have changed, after all - is your evil timeless, Celadroch, or were you waiting for me?",
                 "HydraWinterTiamatCharacter": "You are indeed ancient and terrible. Almost as much as I am.",
-				"FutureTiamatCharacter": "This mockery of the great beast makes me sad. Let us end this."
+                "FutureTiamatCharacter": "This mockery of the great beast makes me sad. Let us end this."
             },
-            "complexity": 3
+            "complexity": 3,
+            "tokenPools": [
+                {
+                    "identifier": "LadyOfTheWoodElementPool",
+                    "name": "Lady of the Wood's Element Pool",
+                    "initialValue": 0,
+                    "minimumValue": 0,
+                    "labelTopColor": "e04b26",
+                    "labelBottomColor": "e04b26"
+                }
+            ]
         }
     ]
 }
